Remove underscored imports
